refactor: AppHeader에 TopSection 통합
