+++ src/components/AppHeader/index.module.css
@@ -1,9 +1,5 @@
 .container {
   display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0 20px;
-  height: 60px;
-  background-color: #333;
-  color: #fff;
+  flex-direction: column;
+  gap: 40px;
 }
