+++ src/components/AppHeader/title.module.css
@@ -1,0 +1,16 @@
+.container {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.logo {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.date {
+  color: var(--text-text-default, #5F6E76);
+
+}
